devops/proyecto_asir/auth-service-lista-cerrada/requirements.txt
Flask==3.0.0
Flask-Cors==4.0.0
PyJWT==2.8.0
Werkzeug==3.0.1
itsdangerous==2.1.2
gunicorn==21.2.0
